Moved files and changed parts of java files to accomodate
