+++ app/src/main/java/ca/unb/mobiledev/projectcs2063/StepsFragment.java
@@ -10,7 +10,7 @@ import androidx.fragment.app.Fragment;
 
 
 
-import static ca.unb.mobiledev.projectcs2063.R.menu.fragment_steps;
+import static ca.unb.mobiledev.projectcs2063.R.layout.fragment_steps;
 
 public class StepsFragment extends Fragment {
     private final static String TAG = "INFO Steps Fragment";
